+++ rider/src/main/kotlin/com/jetbrains/rider/plugins/unity/run/configurations/unityExe/UnityExeConfigurationType.kt
@@ -3,7 +3,7 @@ package com.jetbrains.rider.plugins.unity.run.configurations.unityExe
 import com.intellij.execution.configurations.ConfigurationTypeBase
 import icons.UnityIcons
 
-class UnityExeConfigurationType : ConfigurationTypeBase("RunUnityExe", "Standalone Player", // "RunUnityExe" preserved for compatibility
+class UnityExeConfigurationType : ConfigurationTypeBase(id, "Standalone Player", // "RunUnityExe" preserved for compatibility
     "Unity Standalone Player configuration", UnityIcons.RunConfigurations.UnityExe) {
 
     val factory: UnityExeConfigurationFactory = UnityExeConfigurationFactory(this)
@@ -12,5 +12,9 @@ class UnityExeConfigurationType : ConfigurationTypeBase("RunUnityExe", "Standalo
 
     init {
         addFactory(factory)
+    }
+
+    companion object {
+        const val id = "RunUnityExe"
     }
 }