+++ CHANGELOG.md
@@ -12,18 +12,27 @@ Since 2018.1, the version numbers and release cycle match Rider's versions and r
 * [Commits](https://github.com/JetBrains/resharper-unity/compare/net211...net212)
 * [Milestone](https://github.com/JetBrains/resharper-unity/milestone/45?closed=1)
 
+### Added
+
+- Rider: Warn if trying to commit with unsaved scenes ([#1077](https://github.com/JetBrains/resharper-unity/issues/1077), [RIDER-60824](https://youtrack.jetbrains.com/issue/RIDER-60824), [#2071](https://github.com/JetBrains/resharper-unity/pull/2071))
+- Rider: Support debugging local UWP players ([RIDER-62707](https://youtrack.jetbrains.com/issue/RIDER-62707), [#2098](https://github.com/JetBrains/resharper-unity/pull/2098))
+
 ### Changed
 
 - Methods marked with `[UnitySetUp]` are treated as in use ([#2048](https://github.com/JetBrains/resharper-unity/issues/2048), [#2047](https://github.com/JetBrains/resharper-unity/pull/2047))
+- Rider: Group Unity run configurations in the run config dialog ([#2081](https://github.com/JetBrains/resharper-unity/pull/2081))
+- Rider: Ignore "break on unhandled exception" setting for IL2CPP players ([RIDER-62321](https://youtrack.jetbrains.com/issue/RIDER-62321), [#2098](https://github.com/JetBrains/resharper-unity/pull/2098))
+- Rider: Show localised external Unity documentation if English isn't available ([RIDER-55737](https://youtrack.jetbrains.com/issue/RIDER-55737), [#2050](https://github.com/JetBrains/resharper-unity/pull/2050))
 
 ### Fixed
 
+- Show editor file templates in package folders ([#2090](https://github.com/JetBrains/resharper-unity/pull/2090))
 - Rider: Hide run marker gutter icons for static methods in Unity projects ([RIDER-55734](https://youtrack.jetbrains.com/issue/RIDER-55734), [#2063](https://github.com/JetBrains/resharper-unity/pull/2063))
 - Rider: Error in Unity tool window log is not cleared ([RIDER-59689](https://youtrack.jetbrains.com/issue/RIDER-59689), [#2051](https://github.com/JetBrains/resharper-unity/pull/2051))
 - Rider: External Documentation action now works with Unity API ([#2050](https://github.com/JetBrains/resharper-unity/pull/2050))
 - Rider: Fix wrong page opening in external documentation ([RIDER-57745](https://youtrack.jetbrains.com/issue/RIDER-57745), [#2050](https://github.com/JetBrains/resharper-unity/pull/2050))
 - Rider: External documentation for method parameter now navigates to documentation for method ([https://youtrack.jetbrains.com/issue/RIDER-60297](https://youtrack.jetbrains.com/issue/RIDER-60297), [#2050](https://github.com/JetBrains/resharper-unity/pull/2050))
-- Rider: Show localised external Unity documentation if English isn't available ([RIDER-55737](https://youtrack.jetbrains.com/issue/RIDER-55737), [#2050](https://github.com/JetBrains/resharper-unity/pull/2050))
+- Rider: Fix packages occasionally not showing or updating in Unity Explorer when Unity is not running ([RIDER-62558](https://youtrack.jetbrains.com/issue/RIDER-62558), [#2085](https://github.com/JetBrains/resharper-unity/pull/2085))
 
 
 
@@ -33,8 +42,11 @@ Since 2018.1, the version numbers and release cycle match Rider's versions and r
 
 ### Fixed
 
+- Rider: Remove Unity built in package folders from Perforce content roots ([RIDER-61551](https://youtrack.jetbrains.com/issue/RIDER-61551))
 - Rider: Only add Unity folders to index for Unity packages ([#2076](https://github.com/JetBrains/resharper-unity/pull/2076))
 - Rider: Clean up indexed folders added in previous versions ([#2076](https://github.com/JetBrains/resharper-unity/pull/2076))
+- Rider: Fix unresolved git packages when the `packages-lock.json` is disabled ([#2084](https://github.com/JetBrains/resharper-unity/pull/2084))
+- Rider: Fix unresolved or incorrectly resolved package used as a dependency when `packages-lock.json` is disabled ([#2084](https://github.com/JetBrains/resharper-unity/pull/2084))